requirements_webapp.txt
fastapi
python-multipart
uvicorn
yt-dlp
faster-whisper
google-generativeai
selenium
webdriver-manager
